backend/requirements.txt
# Backend Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets>=12.0,<14.0

# LangChain & LangGraph
langchain==0.3.14
langgraph==0.2.62
langchain-openai==0.3.0
langchain-groq==0.2.3

# Web Search
tavily-python==0.5.0
langchain-community==0.3.14

# Audio
deepgram-sdk==3.8.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# TTS Fallback
edge-tts==6.1.12
httpx>=0.28.0
